third_party/blink/web_tests/external/wpt/css/css-contain/content-visibility/content-visibility-style-containment-001-expected.txt

This is a testharness.js-based test.
[FAIL] switching content-visibility from visible to hidden
  assert_true: expected true got false
[FAIL] switching content-visibility from hidden to visible
  assert_false: expected false got true
[FAIL] switching content-visibility from visible to auto
  assert_true: expected true got false
[FAIL] switching content-visibility from auto to visible
  assert_false: expected false got true
Harness: the test ran to completion.
